% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_plot.R
\name{pair_plot}
\alias{pair_plot}
\title{Pair Plot}
\usage{
pair_plot(data)
}
\arguments{
\item{data}{A data.frame containing numeric variables}
}
\value{
Produces a scatterplot matrix as a side effect
}
\description{
Creates a matrix of scatterplots for all numeric variables in a data frame.
}
\examples{
pair_plot(mtcars)

}
