% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot_pdf.R
\name{save_plot_pdf}
\alias{save_plot_pdf}
\title{Save a ggplot2 Plot to PDF}
\usage{
save_plot_pdf(plot, filename, width = 7, height = 5)
}
\arguments{
\item{plot}{A ggplot object}

\item{filename}{Name of the PDF file to save}

\item{width}{Width of the PDF (in inches)}

\item{height}{Height of the PDF (in inches)}
}
\value{
Invisibly returns the filename
}
\description{
Saves a ggplot2 plot object to a PDF file.
}
\examples{
p <- scatter_with_lm(mtcars, "wt", "mpg")
save_plot_pdf(p, "myplot.pdf", width = 6, height = 4)

}
