\name{Scatterplot}
\alias{Plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Scatterplot
}
\description{
Given data, create a scatterplot with grid and red regression line
}
\usage{
Scatterplot(x,y)
}
\arguments{
  \item{x}{
data for x axis
}
  \item{y}{
data for y axis
}
}
\author{
Zhiyu Shen
}
