% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_with_lm.R
\name{scatter_with_lm}
\alias{scatter_with_lm}
\title{Scatterplot with Regression Line}
\usage{
scatter_with_lm(data, x, y)
}
\arguments{
\item{data}{A data.frame}

\item{x}{Name of the x variable (string)}

\item{y}{Name of the y variable (string)}
}
\value{
A ggplot2 object
}
\description{
Creates a ggplot2 scatterplot with an added linear regression line.
}
\examples{
scatter_with_lm(mtcars, "wt", "mpg")

}
