% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Create a Scatterplot}
\usage{
scatterplot(data, x, y, colour = "blue", na.rm = FALSE)
}
\arguments{
\item{data}{Specify the dataset to use for plot. This argument should be a data.frame. If not already a data.frame, will be converted to one by fortify().}

\item{x}{Specify which column is used as x-axis. The values in this column will be used as the x-values of the points showed in the scatterplot. This argument should be a name of column, without quotes.}

\item{y}{Specify which column is used as y-axis. The values in this column will be used as the y-values of the points showed in the scatterplot. This argument should be a name of column, without quotes.}

\item{colour}{An optional argument, specifying what colour the points will be. This argument should be the name of a colour, with quotes. The default colour is blue.}

\item{na.rm}{An optional argument. If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}
}
\value{
A scatterplot which makes the relationship between two variables specified in the input of the function (x and y) visualized.
}
\description{
Create a Scatterplot
}
\examples{
test <- data.frame(
   x = c(2,4,5,8,10),
   y = c(1,2,3,5,5),
   z = c(1,2,5,7,9))
scatterplot(test, x, y)
scatterplot(test, y, x, "red")

}
